+++ app/code/Magento/InventoryCatalogAdminUi/etc/adminhtml/di.xml
@@ -14,10 +14,6 @@
                     <item name="class" xsi:type="string">Magento\InventoryCatalogAdminUi\Ui\DataProvider\Product\Form\Modifier\SourceItems</item>
                     <item name="sortOrder" xsi:type="number">200</item>
                 </item>
-                <item name="salable_quantity" xsi:type="array">
-                    <item name="class" xsi:type="string">Magento\InventoryCatalogAdminUi\Ui\DataProvider\Product\Form\Modifier\SalableQuantity</item>
-                    <item name="sortOrder" xsi:type="number">300</item>
-                </item>
                 <item name="quantity" xsi:type="array">
                     <item name="class" xsi:type="string">Magento\InventoryCatalogAdminUi\Ui\DataProvider\Product\Form\Modifier\Quantity</item>
                     <item name="sortOrder" xsi:type="number">400</item>
