+++ app/code/Magento/InventoryCatalogAdminUi/view/adminhtml/ui_component/product_form.xml
@@ -210,24 +210,4 @@
             </insertListing>
         </modal>
     </fieldset>
-    <fieldset name="salable_quantity" sortOrder="55">
-        <settings>
-            <label translate="true">Product Salable Quantity</label>
-            <collapsible>true</collapsible>
-            <opened>false</opened>
-            <visible>0</visible>
-        </settings>
-        <container name="salable_quantity_container" template="Magento_InventoryCatalogAdminUi/product/form/salable-quantity" sortOrder="10">
-            <argument name="data" xsi:type="array">
-                <item name="config" xsi:type="array">
-                    <item name="formElement" xsi:type="string">container</item>
-                    <item name="breakLine" xsi:type="boolean">false</item>
-                    <item name="visible" xsi:type="boolean">true</item>
-                    <item name="imports" xsi:type="array">
-                        <item name="salable_quantity" xsi:type="string">${ $.provider }:data.salable_quantity</item>
-                    </item>
-                </item>
-            </argument>
-        </container>
-    </fieldset>
 </form>
