+++ app/code/Magento/InventorySales/Test/Integration/IsProductSalableForRequestedQty/BackorderConditionTest.php
@@ -76,7 +76,7 @@ class BackorderConditionTest extends TestCase
      * @magentoDataFixture ../../../../app/code/Magento/InventoryApi/Test/_files/stocks.php
      * @magentoDataFixture ../../../../app/code/Magento/InventoryApi/Test/_files/stock_source_links.php
      * @magentoDataFixture ../../../../app/code/Magento/InventoryApi/Test/_files/source_items.php
-     * 
+     *
      * @magentoDbIsolation disabled
      */
     public function testBackorderedZeroQtyProductIsSalable()
