+++ app/code/Magento/InventorySalesAdminUi/Model/GetSalableQuantityDataBySku.php
@@ -5,7 +5,7 @@
  */
 declare(strict_types=1);
 
-namespace Magento\InventoryCatalogAdminUi\Model;
+namespace Magento\InventorySalesAdminUi\Model;
 
 use Magento\Inventory\Model\ResourceModel\GetAssignedStockIdsBySku;
 use Magento\InventoryApi\Api\StockRepositoryInterface;
