+++ app/code/Magento/InventorySalesAdminUi/view/adminhtml/web/js/product/grid/cell/salable-quantity.js
@@ -9,7 +9,7 @@ define([
 
     return Column.extend({
         defaults: {
-            bodyTmpl: 'Magento_InventoryCatalogAdminUi/product/grid/cell/salable-quantity.html'
+            bodyTmpl: 'Magento_InventorySalesAdminUi/product/grid/cell/salable-quantity.html'
         },
 
         /**
