+++ app/code/Magento/InventorySalesAdminUi/view/adminhtml/web/template/product/form/salable-quantity.html
