+++ app/code/Magento/InventorySalesAdminUi/view/adminhtml/web/template/product/grid/cell/salable-quantity.html
